[inetstack] Enhancement: Move hashttlcache out
