--- src/rust/inetstack/collections/mod.rs
@@ -1,6 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-pub mod hashttlcache;
-
-pub use hashttlcache::HashTtlCache;
